# runtime/openapi.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "PFun CMA Model Backend"
    },
    "schemes": [
        "https"
    ],
    "paths": {
        "/sdk": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "200 response",
                        "schema": {
                            "$ref": "#/definitions/Empty"
                        }
                    }
                },
                "x-amazon-apigateway-integration": {
                    "responses": {
                        "default": {
                            "statusCode": "200"
                        }
                    },
                    "uri": "arn:{partition}:apigateway:{region_name}:lambda:path/2015-03-31/functions/{api_handler_lambda_arn}/invocations",
                    "passthroughBehavior": "when_no_match",
                    "httpMethod": "POST",
                    "contentHandling": "CONVERT_TO_TEXT",
                    "type": "aws_proxy"
                },
                "security": [
                    {
                        "fake_auth": []
                    }
                ]
            },
            "options": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "200 response",
                        "schema": {
                            "$ref": "#/definitions/Empty"
                        },
                        "headers": {
                            "Access-Control-Allow-Methods": {
                                "type": "string"
                            },
                            "Access-Control-Allow-Origin": {
                                "type": "string"
                            },
                            "Access-Control-Allow-Headers": {
                                "type": "string"
                            },
                            "Access-Control-Expose-Headers": {
                                "type": "string"
                            },
                            "Access-Control-Max-Age": {
                                "type": "string"
                            },
                            "Access-Control-Allow-Credentials": {
                                "type": "string"
                            }
                        }
                    }
                },
                "x-amazon-apigateway-integration": {
                    "responses": {
                        "default": {
                            "statusCode": "200",
                            "responseParameters": {
                                "method.response.header.Access-Control-Allow-Methods": "'GET,OPTIONS'",
                                "method.response.header.Access-Control-Allow-Origin": "'pfun-cma-model-api.p.rapidapi.com'",
                                "method.response.header.Access-Control-Allow-Headers": "'Authorization,Content-Type,X-API-Key,X-Amz-Date,X-Amz-Security-Token,X-Api-Key,X-RapidAPI-Host,X-RapidAPI-Key,X-RapidAPI-Proxy-Secret'",
                                "method.response.header.Access-Control-Expose-Headers": "'X-RapidAPI-Key,X-RapidAPI-Proxy-Secret,X-RapidAPI-Host,X-API-Key,Authorization'",
                                "method.response.header.Access-Control-Max-Age": "'300'",
                                "method.response.header.Access-Control-Allow-Credentials": "'true'"
                            }
                        }
                    },
                    "requestTemplates": {
                        "application/json": "{\"statusCode\": 200}"
                    },
                    "passthroughBehavior": "when_no_match",
                    "type": "mock",
                    "contentHandling": "CONVERT_TO_TEXT"
                }
            }
        },
...
        "/log": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "200 response",
                        "schema": {
                            "$ref": "#/definitions/Empty"
                        }
                    }
                },
                "x-amazon-apigateway-integration": {
                    "responses": {
                        "default": {
                            "statusCode": "200"
                        }
                    },
                    "uri": "arn:{partition}:apigateway:{region_name}:lambda:path/2015-03-31/functions/{api_handler_lambda_arn}/invocations",
                    "passthroughBehavior": "when_no_match",
                    "httpMethod": "POST",
                    "contentHandling": "CONVERT_TO_TEXT",
                    "type": "aws_proxy"
                },
                "security": [
                    {
                        "fake_auth": []
                    }
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "200 response",
                        "schema": {
                            "$ref": "#/definitions/Empty"
                        }
                    }
                },
                "x-amazon-apigateway-integration": {
                    "responses": {
                        "default": {
                            "statusCode": "200"
                        }
                    },
                    "uri": "arn:{partition}:apigateway:{region_name}:lambda:path/2015-03-31/functions/{api_handler_lambda_arn}/invocations",
                    "passthroughBehavior": "when_no_match",
                    "httpMethod": "POST",
                    "contentHandling": "CONVERT_TO_TEXT",
                    "type": "aws_proxy"
                },
                "security": [
                    {
                        "fake_auth": []
                    }
                ]
            },
            "options": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "200 response",
                        "schema": {
                            "$ref": "#/definitions/Empty"
                        },
                        "headers": {
                            "Access-Control-Allow-Methods": {
                                "type": "string"
                            },
                            "Access-Control-Allow-Origin": {
                                "type": "string"
                            },
                            "Access-Control-Allow-Headers": {
                                "type": "string"
                            },
                            "Access-Control-Expose-Headers": {
                                "type": "string"
                            },
                            "Access-Control-Max-Age": {
                                "type": "string"
                            },
                            "Access-Control-Allow-Credentials": {
                                "type": "string"
                            }
                        }
                    }
                },
                "x-amazon-apigateway-integration": {
                    "responses": {
                        "default": {
                            "statusCode": "200",
                            "responseParameters": {
                                "method.response.header.Access-Control-Allow-Methods": "'GET,POST,OPTIONS'",
                                "method.response.header.Access-Control-Allow-Origin": "'pfun-cma-model-api.p.rapidapi.com'",
                                "method.response.header.Access-Control-Allow-Headers": "'Authorization,Content-Type,X-API-Key,X-Amz-Date,X-Amz-Security-Token,X-Api-Key,X-RapidAPI-Host,X-RapidAPI-Key,X-RapidAPI-Proxy-Secret'",
                                "method.response.header.Access-Control-Expose-Headers": "'X-RapidAPI-Key,X-RapidAPI-Proxy-Secret,X-RapidAPI-Host,X-API-Key,Authorization'",
                                "method.response.header.Access-Control-Max-Age": "'300'",
                                "method.response.header.Access-Control-Allow-Credentials": "'true'"
                            }
                        }
                    },
                    "requestTemplates": {
                        "application/json": "{\"statusCode\": 200}"
                    },
                    "passthroughBehavior": "when_no_match",
                    "type": "mock",
                    "contentHandling": "CONVERT_TO_TEXT"
                }
            }
        },
...
    },
    "definitions": {
        "Empty": {
            "type": "object",
            "title": "Empty Schema"
        }
    },
...
    "securityDefinitions": {
        "fake_auth": {
            "in": "header",
            "type": "apiKey",
            "name": "Authorization",
            "x-amazon-apigateway-authtype": "custom",
            "x-amazon-apigateway-authorizer": {
                "type": "token",
                "authorizerUri": "arn:{partition}:apigateway:{region_name}:lambda:path/2015-03-31/functions/{fake_auth_lambda_arn}/invocations"
            }
        }
    }
}